// types.ts
export interface KarabinerRules {
  description?: string
  manipulators?: Manipulator[]
}

export interface Manipulator {
  description?: string
  type: 'basic'
  from: From
  to?: To[]
  to_after_key_up?: To[]
  to_if_alone?: To[]
  to_if_held_down?: To[]
  parameters?: Parameters
  conditions?: Conditions[]
}

export interface Parameters {
  'basic.simultaneous_threshold_milliseconds'?: number
  'basic.to_if_alone_timeout_milliseconds'?: number
  'basic.to_if_held_down_threshold_milliseconds'?: number
}

export type Conditions =
  | FrontMostApplicationCondition
  | DeviceCondition
  | KeybaordTypeCondition
  | InputSourceCondition
  | VaribaleCondition
  | EventChangedCondition
// ...
export interface From {
  key_code?: KeyCode
  simultaneous?: SimultaneousFrom[]
  simultaneous_options?: SimultaneousOptions
  modifiers?: Modifiers
  pointing_button?: 'button1' | 'button2' | 'button3' | 'button4' | 'button5'
}

export interface Modifiers {
  optional?: string[]
  mandatory?: string[]
}

export interface To {
  key_code?: KeyCode
  pointing_button?: 'button1' | 'button2' | 'button3' | 'button4' | 'button5'
  modifiers?: string[]
  hold_down_milliseconds?: number
  shell_command?: string
  set_variable?: {
    name: string
    value: boolean | number | string
  }
  mouse_key?: MouseKey
}